min_requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==1.10.7
pypdf2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
huggingface-hub==0.20.3
numpy==1.24.3
requests==2.28.2
google-generativeai==0.3.1
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.8.4
annoy==1.17.3
scikit-learn==1.2.2
python-magic==0.4.27
pandas==2.3.1
